test: 新增linux环境文件工具类测试
